Extract class with excel reader
